+++ syrup-runtime/src/commonMain/kotlin/io/github/ptitjes/syrup/PluginContext.kt
@@ -7,4 +7,8 @@ interface PluginContext {
     fun <T : Any> contribution(extensionPoint: ExtensionPoint.Singular<T>): LazyDelegate<T>
     fun <T : Any> contributionOrNull(extensionPoint: ExtensionPoint.Singular<T>): LazyDelegate<T?>
     fun <T : Any> contributions(extensionPoint: ExtensionPoint.Plural<T>): LazyDelegate<Set<T>>
+
+    fun <T : Any> sourcedContribution(extensionPoint: ExtensionPoint.Singular<T>): LazyDelegate<Sourced<T>>
+    fun <T : Any> sourcedContributionOrNull(extensionPoint: ExtensionPoint.Singular<T>): LazyDelegate<Sourced<T>?>
+    fun <T : Any> sourcedContributions(extensionPoint: ExtensionPoint.Plural<T>): LazyDelegate<Set<Sourced<T>>>
 }