+++ syrup-runtime/src/commonMain/kotlin/io/github/ptitjes/syrup/Sourced.kt
@@ -1,0 +1,6 @@
+package io.github.ptitjes.syrup
+
+data class Sourced<T : Any>(
+    val source: Plugin,
+    val contribution: T,
+)