+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Next version (unreleased)
 
+## Improvements
+
+- Support retrieving sourced contributions (the contribution along with the source plugin)
+
 # version 0.1.0
 _2026-03-18_